Optimise append stored procedure
Reduced RU consumption of the append stored procedure by reworking the sproc to use a document lookup rather than a query.  In testing this typically reduced the sproc cost by roughly .5-2 units (e.g. writing a single event dropped from 12.3 to 10.2).  RU Usage by sprocs is particularly important as they achieve atomicity by running on the master node of a partition, which means they can start throttling after consuming only a fraction of the available RUs.

To better support gradual roll out / roll back scenarios, the sproc name is now versioned - an 8 character hash is appended to teh name of the sproc based on the contents.  If the sproc contents change, so to will the expected sproc name, which will cause the library to create a new version of the sproc.